Move React Native views into their own package
